O2-O2 cross sections (#70)
Added an OpticalProperty class and a Constituent class for O2-O2
absorption. The classes can be used for other collision-induced
absorption in the future if needed. It currently applies the mole
fraction for oxygen (0.20964) to the total air number density, it could
be updated to be applied to dry air only in the future.

---------

Co-authored-by: Lukas Fehr <lukas.fehr@usask.ca>